Adds missing index.ts to npm package, v1.0.7 (#10)
Adds `index.ts` in npm package to prevent build warning [1] when using the package in projects with  `sourceMaps` enabled. Tested locally with `npm pack` and it was packing the .ts file, should work with `npm publish`.

There might be a better solution by not including `ts.map` files and serving only `.js` and `.d.ts` files but that might need some research.

[1]: `(Emitted value instead of an instance of Error) Cannot find source file 'packages/index.ts': Error: Can't resolve '../index.ts' in...`
To sync with @toi/toix.